fix: [ENG-2522] address PR review findings
- Hot-swap race: clear providerConfigDirty BEFORE awaitAll chain so a task
  arriving during the wait sees a clean flag and skips its inline swap.
  Re-arm only on unexpected throw from the chain.
- Test structure: re-nest 'summary propagation' describe as a sibling of
  'instructionsVar naming' (was misnested under it). 'workspace path
  resolution' and 'runAgentBody / finalize split' now also nest properly
  inside the outer 'FolderPackExecutor' describe.
- Extract shared propagateSummariesUnderLock helper into
  context-tree/propagate-summaries.ts. Curate and folder-pack post-work
  now call it instead of duplicating the dream-lock + diff + propagate
  + manifest block.
- Document the folder-pack Phase 4 task-session timing (session is
  already deleted by executeIterative; folder-pack has no per-task
  sandbox state to preserve through Phase 4).
- Document the query/search exclusion from the awaitProject dispatch
  gate (manifest staleness is absorbed by readManifestIfFresh fallback).
